Move Polymesh types out of staking pallet.
